Use one location service trait, flatten location service folder, remove excessive froms in location service
