+++ graphql/src/schema/queries/mod.rs
@@ -85,7 +85,7 @@ impl Queries {
             Err(error) => return LocationsResponse::Error(error.into()),
         };
 
-        match service_provider.location_query_service.get_locations(
+        match service_provider.location_service.get_locations(
             page.map(PaginationOption::from),
             filter.map(LocationFilter::from),
             convert_sort(sort),
