+++ server/tests/graphql/pagination.rs
@@ -8,7 +8,7 @@ mod graphql {
     use serde_json::json;
     use server::test_utils::setup_all;
     use service::{
-        location::LocationQueryServiceTrait,
+        location::LocationServiceTrait,
         service_provider::{ServiceContext, ServiceProvider},
         ListError, ListResult,
     };
@@ -23,7 +23,7 @@ mod graphql {
 
     struct TestService(pub Box<GetLocations>);
 
-    impl LocationQueryServiceTrait for TestService {
+    impl LocationServiceTrait for TestService {
         fn get_locations(
             &self,
             pagination: Option<PaginationOption>,
@@ -49,7 +49,7 @@ mod graphql {
             connection_manager: StorageConnectionManager,
         ) -> ServiceProvider {
             let mut service_provider = ServiceProvider::new(connection_manager);
-            service_provider.location_query_service = Box::new(self);
+            service_provider.location_service = Box::new(self);
             service_provider
         }
     }
