+++ service/src/location/tests/delete.rs
@@ -11,13 +11,15 @@ mod query {
     };
 
     use crate::{
-        current_store_id, location::delete::DeleteLocationError, service_provider::ServiceProvider,
+        current_store_id,
+        location::delete::{DeleteLocationError, LocationInUse},
+        service_provider::ServiceProvider,
     };
 
     #[actix_rt::test]
-    async fn delete_location_service_errors() {
+    async fn location_service_delete_errors() {
         let (_, _, connection_manager, _) =
-            setup_all("delete_location_service_errors", MockDataInserts::all()).await;
+            setup_all("location_service_delete_errors", MockDataInserts::all()).await;
 
         let connection = connection_manager.connection().unwrap();
         let location_repository = LocationRepository::new(&connection);
@@ -26,7 +28,7 @@ mod query {
 
         let service_provider = ServiceProvider::new(connection_manager);
         let context = service_provider.context().unwrap();
-        let service = service_provider.delete_location_service;
+        let service = service_provider.location_service;
 
         let locations_not_in_store = location_repository
             .query_by_filter(
@@ -68,10 +70,10 @@ mod query {
 
         assert_eq!(
             service.delete_location(DeleteLocation { id: location_id }, &context),
-            Err(DeleteLocationError::LocationInUse {
+            Err(DeleteLocationError::LocationInUse(LocationInUse {
                 stock_lines,
                 invoice_lines
-            })
+            }))
         );
 
         // Location is not empty (stock_lines in use)
@@ -85,22 +87,22 @@ mod query {
 
         assert_eq!(
             service.delete_location(DeleteLocation { id: location_id }, &context),
-            Err(DeleteLocationError::LocationInUse {
+            Err(DeleteLocationError::LocationInUse(LocationInUse {
                 stock_lines,
                 invoice_lines
-            })
+            }))
         );
     }
     #[actix_rt::test]
-    async fn delete_location_service_success() {
+    async fn location_service_delete_success() {
         let (_, _, connection_manager, _) =
-            setup_all("delete_location_service_success", MockDataInserts::all()).await;
+            setup_all("location_service_delete_success", MockDataInserts::all()).await;
 
         let connection = connection_manager.connection().unwrap();
         let location_repository = LocationRepository::new(&connection);
         let service_provider = ServiceProvider::new(connection_manager);
         let context = service_provider.context().unwrap();
-        let service = service_provider.delete_location_service;
+        let service = service_provider.location_service;
 
         assert_eq!(
             service.delete_location(
