+++ service/src/location/tests/insert.rs
@@ -16,7 +16,7 @@ mod query {
         let location_repository = LocationRepository::new(&connection);
         let service_provider = ServiceProvider::new(connection_manager);
         let context = service_provider.context().unwrap();
-        let service = service_provider.insert_location_service;
+        let service = service_provider.location_service;
 
         let locations_in_store = location_repository
             .query_by_filter(
@@ -61,7 +61,7 @@ mod query {
         let location_repository = LocationRepository::new(&connection);
         let service_provider = ServiceProvider::new(connection_manager);
         let context = service_provider.context().unwrap();
-        let service = service_provider.insert_location_service;
+        let service = service_provider.location_service;
 
         let result_location = Location {
             id: "new_id".to_owned(),
