+++ service/src/location/tests/mod.rs
@@ -1,0 +1,8 @@
+#[cfg(test)]
+mod delete;
+#[cfg(test)]
+mod insert;
+#[cfg(test)]
+mod query;
+#[cfg(test)]
+mod update;
