+++ service/src/location/validate.rs
@@ -1,45 +1,30 @@
 use domain::location::LocationFilter;
 use repository::{
-    schema::LocationRow, LocationRepository, LocationRowRepository, StorageConnection,
+    schema::LocationRow, LocationRepository, LocationRowRepository, RepositoryError,
+    StorageConnection,
 };
 
-use crate::{current_store_id, WithDBError};
-
-pub struct LocationWithCodeAlreadyExists;
+use crate::current_store_id;
 
 pub fn check_location_code_is_unique(
     id: &String,
-    code_option: &Option<String>,
+    code: &String,
     connection: &StorageConnection,
-) -> Result<(), WithDBError<LocationWithCodeAlreadyExists>> {
-    if let Some(code) = code_option {
-        let current_store_id = current_store_id(connection)?;
-        let locations = LocationRepository::new(connection).query_by_filter(
-            LocationFilter::new()
-                .code(|f| f.equal_to(code))
-                .id(|f| f.not_equal_to(id))
-                .store_id(|f| f.equal_to(&current_store_id)),
-        )?;
+) -> Result<bool, RepositoryError> {
+    let current_store_id = current_store_id(connection)?;
+    let locations = LocationRepository::new(connection).query_by_filter(
+        LocationFilter::new()
+            .code(|f| f.equal_to(code))
+            .id(|f| f.not_equal_to(id))
+            .store_id(|f| f.equal_to(&current_store_id)),
+    )?;
 
-        if locations.len() > 0 {
-            return Err(WithDBError::Error(LocationWithCodeAlreadyExists {}));
-        }
-    }
-
-    Ok(())
+    Ok(locations.len() == 0)
 }
-
-pub struct LocationDoesNotExist;
 
 pub fn check_location_exists(
     id: &str,
     connection: &StorageConnection,
-) -> Result<LocationRow, WithDBError<LocationDoesNotExist>> {
-    let location_option = LocationRowRepository::new(connection).find_one_by_id(id)?;
-
-    if let Some(location) = location_option {
-        Ok(location)
-    } else {
-        Err(WithDBError::Error(LocationDoesNotExist {}))
-    }
+) -> Result<Option<LocationRow>, RepositoryError> {
+    Ok(LocationRowRepository::new(connection).find_one_by_id(id)?)
 }
