+++ service/src/service_provider.rs
@@ -1,19 +1,10 @@
 use repository::{RepositoryError, StorageConnection, StorageConnectionManager};
 
-use crate::location::{
-    delete::{DeleteLocationService, DeleteLocationServiceTrait},
-    insert::{InsertLocationService, InsertLocationServiceTrait},
-    query::LocationQueryService,
-    update::{UpdateLocationService, UpdateLocationServiceTrait},
-    LocationQueryServiceTrait,
-};
+use crate::location::{LocationServiceTrait, LocationService};
 
 pub struct ServiceProvider {
     pub connection_manager: StorageConnectionManager,
-    pub location_query_service: Box<dyn LocationQueryServiceTrait>,
-    pub insert_location_service: Box<dyn InsertLocationServiceTrait>,
-    pub update_location_service: Box<dyn UpdateLocationServiceTrait>,
-    pub delete_location_service: Box<dyn DeleteLocationServiceTrait>,
+    pub location_service: Box<dyn LocationServiceTrait>,
 }
 
 pub struct ServiceContext {
@@ -24,10 +15,7 @@ impl ServiceProvider {
     pub fn new(connection_manager: StorageConnectionManager) -> Self {
         ServiceProvider {
             connection_manager,
-            location_query_service: Box::new(LocationQueryService {}),
-            insert_location_service: Box::new(InsertLocationService {}),
-            update_location_service: Box::new(UpdateLocationService {}),
-            delete_location_service: Box::new(DeleteLocationService {}),
+            location_service: Box::new(LocationService {}),
         }
     }
