+++ service/src/validate.rs
@@ -1,16 +1,10 @@
-use repository::StorageConnection;
+use repository::{RepositoryError, StorageConnection};
 
-use crate::{current_store_id, WithDBError};
-
-pub struct RecordDoesNotBelongToCurrentStore;
+use crate::current_store_id;
 
 pub fn check_record_belongs_to_current_store(
     store_id: &str,
     connection: &StorageConnection,
-) -> Result<(), WithDBError<RecordDoesNotBelongToCurrentStore>> {
-    if store_id == &current_store_id(connection)? {
-        Ok(())
-    } else {
-        Err(WithDBError::Error(RecordDoesNotBelongToCurrentStore))
-    }
+) -> Result<bool, RepositoryError> {
+    Ok(store_id == &current_store_id(connection)?)
 }
